Moved exersices from Meetings 6 and 10 to common folder
