+++ Exercises/06_wheres_wally_traditional.py
